converting store and root-reducer to ts
