+++ package.json
@@ -10,6 +10,7 @@
     "@types/node": "^20.4.9",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
+    "@types/redux-logger": "^3.0.9",
     "firebase": "^10.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
