+++ src/store/list/list.selector.ts
@@ -1,6 +1,7 @@
 import { ListState } from "./list.reducer";
 import { createSelector } from "reselect";
+import { RootState } from "../store";
 
-export const selectListReducer = (state): ListState => state.list;
+export const selectListReducer = (state: RootState): ListState => state.list;
 
 export const selectList = createSelector(selectListReducer, list => list);
