+++ src/store/root-reducer.ts
@@ -6,6 +6,6 @@ import { userReducer } from "./user/user.reducer";
 import { listReducer } from "./list/list.reducer";
 
 export const rootReducer = combineReducers({
-  user: userReducer,
-  list: listReducer,
+	user: userReducer,
+	list: listReducer,
 });
