+++ src/store/user/user.selector.ts
@@ -1,7 +1,8 @@
 import { UserState } from "./user.reducer";
 import { createSelector } from "reselect";
+import { RootState } from "../store";
 
-export const selectUserReducer = (state): UserState => state.user;
+export const selectUserReducer = (state: RootState): UserState => state.user;
 
 export const selectCurrentUser = createSelector(
 	selectUserReducer,
